[MCUX-62596][ELE] Fix hash algo selection and error cases for RSA sign/verify
For PKCS1_V15, the algorithm should be chosen based on the 'md_alg' parameter,
and for PSS it should be 'ctx->hash_id' or 'mgf1_hash_id' - depends if the _ext() API is used.
For PSS, the MBEDTLS_MD_NONE value should result in a bad argument error.
Fixed typos in comments.