[MCUX-63303][ELE] Use common config file for both ECC opaque key examples
There was a conflict between ECC opaque key example-specific config files
and the common config file within the port layer, which prevented setting the
MBEDTLS_CONFIG_FILE macro without specifying a folder path. This conflict
was resolved by removing the example-specific configs, and the format of the
macro was fixed to only contain the common config's filename.